New Server: Selector Server
A new type of server can now be chosen to be used instead of the Multithreaded Server. For now only a simple Echo Selector Server is working. ComUtils will need to be adapted to be used with unstable data and fragmented packets. For now a simple solution will lead us to use an intermediary buffer on ComUtils to be used as an IO pipe between the Selector Server and the various read commands. Write operations will not be a problem.