Works RandomAI between server and client in automatic mode
Restructurated class Game and update methods to apply AI. In ComUtils: expectedCards saved also in case DRAW (it was necessary for the automatic play mode). Now to connect to the server is required a 0 or a 1 after the connect localhost command, to play in manual (0) or automatic (1) mode. This number should match with the cInteractive of GameData (TODO set cInteractive of GameData to the read value 0 or 1)