Refactored AI to Source Interface
The new AI refactor now allows the Client and Server to use a semi-identical version of the AI. Problems: for some reason, the Server and the Client AI don't agree with the order of commands, and send eachother conflicting commands, leading to a Protocol standstill. Seems like as both start on the INIT stage, they both take the same actions, leading to an eventual failure.