Merge Refactored Artificial Intelligence (#8)
* Added random server IA

* Works RandomAI between server and client in automatic mode

Restructurated class Game and update methods to apply AI. In ComUtils: expectedCards saved also in case DRAW (it was necessary for the automatic play mode). Now to connect to the server is required a 0 or a 1 after the connect localhost command, to play in manual (0) or automatic (1) mode. This number should match with the cInteractive of GameData (TODO set cInteractive of GameData to the read value 0 or 1)

* Added IntelligentAI classes

* Set AI implements Source. TODO URGENT Fix refactor changes

* Refactored AI to Source Interface

The new AI refactor now allows the Client and Server to use a semi-identical version of the AI. Problems: for some reason, the Server and the Client AI don't agree with the order of commands, and send eachother conflicting commands, leading to a Protocol standstill. Seems like as both start on the INIT stage, they both take the same actions, leading to an eventual failure.

* Fixed some parts of the Client

Some other parts are now broken :) :) :)