Still working on POST/PUT REST CRUD
Not working for the most part, only works when breaking the cool embedded data, and thus only showing the pk of the objects instead of the object's field on the API GET calls. We don't want that, cause that means more work on the front end and more API calls. The final objective is passing the airline and airplane PKs to the API, and it automagically searching and replacing those with the real models. It wont fucking work tho.